+++ backend/src/main/resources/db/migration/V12__create_chat_table.sql
@@ -1,8 +1,8 @@
 CREATE TABLE chat
 (
     id BIGSERIAL PRIMARY KEY,
-    sender INT NOT NULL,
-    recipient INT NOT NULL,
+    sender BIGINT NOT NULL,
+    recipient BIGINT NOT NULL,
     message TEXT NOT NULL DEFAULT '',
     timestamp TIMESTAMP NOT NULL
 );
